Organized imports, cleaned up index.tsx, removed unused auto-generated files
Up next: Remove option to add content with no subheading to reduce code complexity in `route-generator.tsx`, `course-data.tsx` and `SideNavigationComponent.tsx`
